+++ TypeContractor/TypeScript/DestinationTypes.cs
@@ -2,6 +2,7 @@ namespace TypeContractor.TypeScript;
 
 public static class DestinationTypes
 {
+    public const string AnyType = "any";
     public const string StringType = "string";
     public const string Boolean = "boolean";
     public const string Number = "number";