+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Add default map from `System.Object` to `any`
+
+### Fixed
+
+- Fix writing dictionaries with custom types as values wrapped inside lists.
+  E.g. converting `Dictionary<Guid, IEnumerable<FormulaDto>>` to
+  `{ [key: string]: FormulaDto[] }`
+- Fix writing dictionaries wrapped inside other dictionaries, e.g. making sure
+  `Dictionary<Guid, Dictionary<string, IEnumerable<FormulaDto>>>` correctly
+  translates to `{ [key: string]: { [key: string]: FormulaDto[] } }`.
+
 ## [0.9.0] - 2024-01-05
 
 ### Added